we started to support new django-select2 versions. #78
